Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-multipart==0.0.6
pydantic==2.5.0
pymongo==4.6.0
python-dotenv==1.0.0
Pillow==10.1.0
groq>=0.30.0
httpx>=0.27.0
qrcode[pil]>=7.4.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
boto3>=1.34.0
twilio>=9.0.0
pyotp>=2.9.0
cryptography>=42.0.0
pydub>=0.25.1
python-telegram-bot>=20.7
